requirements.txt
pandas
transformers
torch
dvc
boto3
fastapi
uvicorn
pyyaml
